fix: use int64 for tar size results to avoid overflow on large contexts
SizeResults inner type, completeBuildContextUpload return type, and
both getFileSize call-sites changed from int to int64. On 32-bit
targets int(getFileSize) would silently truncate tarballs larger than
2 GiB, corrupting _REPO_BUILD_CONTEXT_TAR_SIZES.

+++ src/docker/context_upload.nim
@@ -65,7 +65,7 @@ type
     string,
     OrderedTableRef[
       string,
-      OrderedTableRef[string, int],
+      OrderedTableRef[string, int64],
     ],
   ]
 
@@ -392,7 +392,7 @@ proc completeBuildContextUpload(
     image:       DockerImage,
     snapshot:    ContextSnapshotEntry,
     contextName: string,
-): (string, int, seq[SkippedFile]) =
+): (string, int64, seq[SkippedFile]) =
   ## Complete a single context upload by creating the attestation manifest.
   ## Returns (digest, tarSize, skippedFiles) where digest is the context manifest
   ## digest (without sha256: prefix), tarSize is the tarball size in bytes, and
@@ -452,7 +452,7 @@ proc completeBuildContextUpload(
         "; the tarball may have been tampered with or replaced",
       )
     let
-      tarSize = int(getFileSize(tarPath))
+      tarSize = getFileSize(tarPath)
       layer   = DockerManifest(
         kind:       DockerManifestType.layer,
         name:       image.withDigest(actualHash),
@@ -502,7 +502,7 @@ proc completeBuildContextUpload(
       )
     try:
       let
-        tarSize = int(getFileSize(tarPath))
+        tarSize = getFileSize(tarPath)
         layer   = DockerManifest(
           kind:       DockerManifestType.layer,
           name:       image,
@@ -571,11 +571,11 @@ proc completeBuildContextUploads*(
           buildContexts[image.registry][repoPath][contextName] = attestDigest
           discard sizeResults.hasKeyOrPut(
             image.registry,
-            newOrderedTable[string, OrderedTableRef[string, int]](),
+            newOrderedTable[string, OrderedTableRef[string, int64]](),
           )
           discard sizeResults[image.registry].hasKeyOrPut(
             repoPath,
-            newOrderedTable[string, int](),
+            newOrderedTable[string, int64](),
           )
           sizeResults[image.registry][repoPath][contextName] = tarSize
           # Collect skipped files: disk strategy returns them directly;
